back/requirements.txt
pyngrok
azure-cognitiveservices-speech
flask
flask_cors
transformers
torch